fix NW 2017 clipping parsing int type and its state

+++ Navisworks/NWClippingPlanes.cs
@@ -10,7 +10,7 @@ namespace ARUP.IssueTracker.Navisworks
     {
         public string Type { get; set; }
         public int Version { get; set; }
-        public List<int> Normal { get; set; }
+        public List<Int64> Normal { get; set; }
         public double Distance { get; set; }
         public bool Enabled { get; set; }
     }
